+++ src/test/javascript/spec/component/course/course-scores/course-scores.component.spec.ts
@@ -30,15 +30,10 @@ import { ArtemisTestModule } from '../../../test.module';
 import { GradeType, GradingScale } from 'app/entities/grading-scale.model';
 import { GradingSystemService } from 'app/grading-system/grading-system.service';
 import { GradeStep } from 'app/entities/grade-step.model';
+import { MockTranslateValuesDirective } from '../../../helpers/mocks/directive/mock-translate-values.directive';
 
 chai.use(sinonChai);
 const expect = chai.expect;
-
-// tslint:disable-next-line:directive-selector
-@Directive({ selector: '[translateValues]' })
-export class MockTranslateValuesDirective {
-    @Input('translateValues') data: any;
-}
 
 describe('CourseScoresComponent', () => {
     let fixture: ComponentFixture<CourseScoresComponent>;
