+++ src/test/javascript/spec/component/exam/manage/student-exams/student-exam-detail.component.spec.ts
@@ -22,7 +22,6 @@ import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 import * as chai from 'chai';
 import { ReactiveFormsModule } from '@angular/forms';
-import { MockTranslateValuesDirective } from '../../../course/course-scores/course-scores.component.spec';
 import { Exercise } from 'app/entities/exercise.model';
 import { ModelingExercise, UMLDiagramType } from 'app/entities/modeling-exercise.model';
 import { ExerciseGroup } from 'app/entities/exercise-group.model';
@@ -36,6 +35,7 @@ import { StudentExamDetailTableRowComponent } from 'app/exam/manage/student-exam
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { GradingSystemService } from 'app/grading-system/grading-system.service';
 import { DataTableComponent } from 'app/shared/data-table/data-table.component';
+import { MockTranslateValuesDirective } from '../../../../helpers/mocks/directive/mock-translate-values.directive';
 
 chai.use(sinonChai);
 const expect = chai.expect;
