+++ src/test/javascript/spec/helpers/mocks/directive/mock-translate-values.directive.ts
@@ -1,0 +1,6 @@
+import { Directive, Input } from '@angular/core';
+// tslint:disable-next-line:directive-selector
+@Directive({ selector: '[translateValues]' })
+export class MockTranslateValuesDirective {
+    @Input('translateValues') data: any;
+}
